Move Parameters to workflow common package
